refactor(review): split experiment recommendation logic
